Add support for querying PuppetDB
This makes it so that after receiving a pull_request webhook from
GitHub, the worker will pull the list of changed files from the pull
request and query PuppetDB for all active nodes which contain resources
declared in the `.pp` files from that list. The idea is to use this list
to pick which nodes to compile catalogs for, since it will cover many of
the nodes affected by the pull request.