Add (untested) support for running puppet preview
This adds code that theoretically should run `puppet preview` on the
nodes affected by the PR represented by the received webhook and do some
very basic processing of the results. It's completely untested, and
there's definitely more processing to do, as well as adding
elasticsearch reporting capabilities. In theory, though, it should set
the status on the PR to `success` if all catalogs compiled successfully,
`failure` if any catalogs failed to compile, or `error` if the process
errored out or something threw an exception.