Use armature to deploy environments
This adds [armature](https://github.com/danielparks/armature) support
for deploying branches and pull requests as environments. Now, the tool
will automatically deploy the base branch for the PR and the merged
result of the PR (i.e., GitHub's `refs/pull/###/merge`) as Puppet
environments. It will deploy the base branch as its normal name, and the
PR as `pr_###`, where `###` is the number for the PR.

This also fixes the call to `puppet preview`.

With this commit, the tool works at a basic level to receive a webhook,
run `puppet preview` as appropriate, and update GitHub.